refactoring int to long

+++ SimpleCalculator.Domain/Entities/RegisterInfo.cs
@@ -2,7 +2,7 @@
 {
 	public class RegisterInfo
 	{
-		public int CurrentValue { get; set; }
+		public long CurrentValue { get; set; }
 
 		public Queue<Command> Commands { get; set; } = new Queue<Command>();
 	}
